requirements.txt
langchain==0.1.0
langchain-community==0.0.10
streamlit==1.28.1
arxiv==2.1.0
python-dotenv==1.0.0
requests==2.31.0
beautifulsoup4==4.12.2
pandas==2.1.3
numpy>=1.26.0 